Updating README with auto detection of path
This became one ugly duckling, you are probably getting tired of me @MarvinJanssen but would love some input of this. I was surprised that the were no simpler way to count the number of objects and hard coding "7" as the length is most definitely not suitable.

I've tested and it works but this needs improvement before we can merge.